fix(OpenAPI/EKS/NPList): fixed volume size type

+++ .gen/pipeline/pipeline/model_eks_node_pool_all_of.go
@@ -15,7 +15,7 @@ type EksNodePoolAllOf struct {
 	Autoscaling NodePoolAutoScaling `json:"autoscaling,omitempty"`
 
 	// Size of the EBS volume in GBs of the nodes in the pool.
-	VolumeSize float32 `json:"volumeSize,omitempty"`
+	VolumeSize int32 `json:"volumeSize,omitempty"`
 
 	// Machine instance type.
 	InstanceType string `json:"instanceType"`
