+++ .gen/pipeline/pipeline/model_node_pool.go
@@ -24,7 +24,7 @@ type NodePool struct {
 	Autoscaling NodePoolAutoScaling `json:"autoscaling,omitempty"`
 
 	// Size of the EBS volume in GBs of the nodes in the pool.
-	VolumeSize float32 `json:"volumeSize,omitempty"`
+	VolumeSize int32 `json:"volumeSize,omitempty"`
 
 	// Machine instance type.
 	InstanceType string `json:"instanceType"`
